the cat is a lie
fait les modifs suggérées dans #25
